Refactoring of Tabular Dataset code (MLDB-1507)
