+++ plugins/column_types.h
@@ -1,6 +1,6 @@
 /** column_types.h                                       -*- C++ -*-
     Jeremy Barnes, 27 March 2016
-    Copyright (c) 2016 Datacratic Inc.  All rights reserved.
+    This file is part of MLDB. Copyright 2016 Datacratic. All rights reserved.
 
     Structure to record types and ranges of values in a column to aid
     with determining its size.
@@ -8,104 +8,33 @@
 
 #pragma once
 
-#include "mldb/sql/cell_value.h"
+#include <memory>
 
 namespace Datacratic {
 namespace MLDB {
 
+struct CellValue;
+struct ExpressionValueInfo;
 
 /*****************************************************************************/
 /* COLUMN TYPES                                                              */
 /*****************************************************************************/
    
+/** This is an accumulator that keeps statistics on the types of values that
+    a column could have.  It's useful for knowing how to treat a column in
+    an algorithm.
+*/
+
 struct ColumnTypes {
-    ColumnTypes()
-        : numNulls(false), numIntegers(false),
-          minNegativeInteger(0), maxPositiveInteger(0),
-          numReals(false), numStrings(false), numBlobs(false),
-          numOther(false)
-    {
-    }
+    ColumnTypes();
 
-    void update(const CellValue & val)
-    {
-        // Record the type
-        switch (val.cellType()) {
-        case CellValue::EMPTY:
-            numNulls += 1;  break;
-        case CellValue::FLOAT:
-            numReals += 1;  break;
+    void update(const CellValue & val);
 
-        case CellValue::INTEGER:
-            numIntegers += 1;
-            if (val.isUInt64()) {
-                maxPositiveInteger = std::max(maxPositiveInteger, val.toUInt());
-            }
-            else {
-                minNegativeInteger = std::min(minNegativeInteger, val.toInt());
-            }
-            break;
-        case CellValue::ASCII_STRING:
-        case CellValue::UTF8_STRING:
-            numStrings += 1;  break;
-        case CellValue::BLOB:
-            numBlobs += 1;  break;
-        default:
-            numOther += 1;  break;
-        }
-    }
-
-    void update(const ColumnTypes & other)
-    {
-        numNulls = numNulls + other.numNulls;
-        numIntegers = numIntegers + other.numIntegers;
-        minNegativeInteger
-            = std::min(minNegativeInteger, other.minNegativeInteger);
-        maxPositiveInteger
-            = std::max(maxPositiveInteger, other.maxPositiveInteger);
-        numReals = numReals + other.numReals;
-        numStrings = numStrings + other.numStrings;
-        numBlobs = numBlobs + other.numBlobs;
-        numOther = numOther + other.numOther;
-    }
+    void update(const ColumnTypes & other);
 
     std::shared_ptr<ExpressionValueInfo>
-    getExpressionValueInfo() const
-    {
-        if (!numNulls && !numReals && !numStrings && !numOther) {
-            // Integers only
-            if (minNegativeInteger == 0) {
-                // All positive
-                return std::make_shared<Uint64ValueInfo>();
-            }
-            else if (maxPositiveInteger <= (1ULL << 63)) {
-                // Fits in a 64 bit integer
-                return std::make_shared<IntegerValueInfo>();
-            }
-            else {
-                // Out of range of either positive or negative integers
-                // only.  We say it's an atom.
-                return std::make_shared<AtomValueInfo>();
-            }
-        }
-        else if (!numNulls && !numStrings && !numOther) {
-            // Reals and integers.  If all integers are representable as
-            // doubles, in other words a maximum of 53 bits, then we're all
-            // doubles.
-            if (maxPositiveInteger < (1ULL << 53)
-                && minNegativeInteger > -(1LL << 53)) {
-                return std::make_shared<Float64ValueInfo>();
-            }
-            // Doubles would lose precision.  It's an atom.
-            return std::make_shared<AtomValueInfo>();
-        }
-        else if (!numNulls && !numIntegers && !numReals && !numOther) {
-            return std::make_shared<Utf8StringValueInfo>();
-        }
-        else {
-            return std::make_shared<AtomValueInfo>();
-        }
-    }
+    getExpressionValueInfo() const;
+
     uint64_t numNulls;
 
     uint64_t numIntegers;
